requirements.txt
alembic==1.4.2
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
crypto==1.4.1
cryptography==2.9.2
dnspython==2.0.0
email-validator==1.1.1
Flask==1.1.2
Flask-Login==0.5.0
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.3
Flask-WTF==0.14.3
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
marshmallow==3.7.0
marshmallow-dataclass==7.6.0
mypy-extensions==0.4.3
Naked==0.1.31
numpy==1.19.0
pandas==1.0.5
psycopg2==2.7.7
psycopg2-binary==2.8.5
py-bcrypt==0.4
pycparser==2.20
pycryptodome==3.4.3
PyJWT==1.7.1
pyOpenSSL==19.1.0
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
shellescape==3.8.1
six==1.15.0
SQLAlchemy==1.3.18
typing-extensions==3.7.4.2
typing-inspect==0.6.0
urllib3==1.25.9
Werkzeug==1.0.1
WTForms==2.3.1
WTForms-JSON==0.3.3
